+++ CHANGELOG.md
@@ -9,6 +9,8 @@
         - removing debug info for `-r` introduced in a0f6b6e. [#180](https://github.com/shenwei356/seqkit/issues/180)
     - `seqkit fx2tab`:
         - new flag `-C/--base-count` for counting bases.
+    - `seqkit subseq`:
+        - fix bug for BED with empty columns. [#195](https://github.com/shenwei356/seqkit/issues/195)
     - `seqkit genautocomplete`: 
         - support bash|zsh|fish|powershell.
 - [SeqKit v0.15.0](https://github.com/shenwei356/seqkit/releases/tag/v0.15.0)